third_party/WebKit/LayoutTests/bluetooth/requestDevice/name-missing-device.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => {
      assert_equals(device.name, 'hmm'); // TODO SHOULD FAIL?
      return device.gatt.connect();
    })
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => characteristic.readValue())
    .then(value => {
      let decoder = new TextDecoder('utf-8');
      let value_str = decoder.decode(value);
      assert_equals(value_str, '');
    });
}, 'An unnamed device can be obtained by advertised service UUID.');

promise_test(t => {
  return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
  .then(() => promise_rejects(
    t, 'NotFoundError', requestDeviceWithKeyDown({
      filters: [{
        name: '',
        services: ['heart_rate']
      }]})));
}, 'An unnamed device can not be obtained by empty name filter.');

promise_test(t => {
  return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
  .then(() => promise_rejects(
    t, 'NotFoundError', requestDeviceWithKeyDown({
      filters: [{
        name: 'a',
        services: ['heart_rate']
      }]})));
}, 'An unnamed device can not be obtained by name filter.');

</script>
